# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MailboxConnectionCompleteRequest": {
        "additionalProperties": false,
        "description": "Only the state and code. There is deliberately no mailbox field.\n\nThe mailbox is taken from the stored attempt the state identifies. Because\nthe contract forbids unknown fields, a callback that tries to name a mailbox\nis rejected before it reaches the service.",
        "properties": {
          "code": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "state": {
            "maxLength": 512,
            "minLength": 16,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "code"
        ],
        "title": "MailboxConnectionCompleteRequest",
        "type": "object"
      },
      "MailboxConnectionOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Code",
            "type": "string"
          },
          "mailbox_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailbox Id"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "code"
        ],
        "title": "MailboxConnectionOutcomeResponse",
        "type": "object"
      },
      "MailboxConnectionStartResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "maxLength": 80,
            "minLength": 3,
            "title": "Attempt Id",
            "type": "string"
          },
          "authorization_url": {
            "maxLength": 4096,
            "minLength": 20,
            "title": "Authorization Url",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "mailbox_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9-]{2,79}$",
            "title": "Mailbox Id",
            "type": "string"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "mailbox_id",
          "attempt_id",
          "authorization_url",
          "expires_at"
        ],
        "title": "MailboxConnectionStartResponse",
        "type": "object"
      },
...
      "RefusalResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Code",
            "type": "string"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "RefusalResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Private service contract for grounded support-ticket triage.",
    "title": "ResolveFlow AI Internal API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/mailboxes/connection/complete": {
      "post": {
        "operationId": "complete_connection_v1_mailboxes_connection_complete_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailboxConnectionCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxConnectionOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxConnectionOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Complete Connection",
        "tags": [
          "mailboxes"
        ]
      }
    },
    "/v1/mailboxes/{mailbox_id}/connection": {
      "delete": {
        "operationId": "revoke_connection_v1_mailboxes__mailbox_id__connection_delete",
        "parameters": [
          {
            "in": "path",
            "name": "mailbox_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{2,79}$",
              "title": "Mailbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxConnectionOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxConnectionOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Revoke Connection",
        "tags": [
          "mailboxes"
        ]
      }
    },
    "/v1/mailboxes/{mailbox_id}/connection/start": {
      "post": {
        "operationId": "start_connection_v1_mailboxes__mailbox_id__connection_start_post",
        "parameters": [
          {
            "in": "path",
            "name": "mailbox_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{2,79}$",
              "title": "Mailbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxConnectionStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxConnectionOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Start Connection",
        "tags": [
          "mailboxes"
        ]
      }
    },
...
  }
}